fix: implicit numeric widening in Work.scala

+++ sdk/src/main/scala/ec/iex/core/Work.scala
@@ -28,9 +28,11 @@ object WorkState extends Enumeration {
 
 object Work {
 
+  import ec.iex.util._
+
   def buyWorkOrder(iexecHubContract: IexecHub, workerPoolId: Int, workerPoolAddress: String, appAddress: String, datasetAddress: String, parameters: String, callbackAddress: String, beneficiaryAddress: String) = {
 
-    val txReceipt = iexecHubContract.buyForWorkOrder(BigInteger.valueOf(workerPoolId), workerPoolAddress, appAddress, datasetAddress, parameters, callbackAddress, beneficiaryAddress).send()
+    val txReceipt = iexecHubContract.buyForWorkOrder(workerPoolId, workerPoolAddress, appAddress, datasetAddress, parameters, callbackAddress, beneficiaryAddress).send()
 
     val workId = "0x" + txReceipt.getLogs().get(1).getTopics().get(1).replace("0x", "").replaceFirst("^0+(?!$)", "")
